cross validation

+++ run_plasma.py
@@ -199,7 +199,7 @@ def run_plasma(name, data_dir, params_path, filter_path, cluster_map_path, barco
     # cumu_int = floors + adds + (1 - np.roll(adds, 1))
     # allocs = np.copy(cumu_int)
     # allocs[1:] -= cumu_int[:-1]
-    allocs = allocs_raw + adds + (1 - np.roll(adds, 1))
+    allocs = (allocs_raw + adds + (1 - np.roll(adds, 1))).astype(int)
     part_idxs = np.concatenate([np.full(val, ind) for ind, val in enumerate(allocs)])
     partitions = np.random.permutation(part_idxs)
     print(rems) ####
